server: Move `polar.benefit.service.*` to `polar.benefit.service` or `polar.benefit.grant.service`
