+++ server/polar/benefit/service.py
@@ -31,9 +31,9 @@ from polar.redis import Redis
 from polar.webhook.service import webhook as webhook_service
 from polar.worker import enqueue_job
 
-from ..schemas import BenefitCreate, BenefitUpdate
-from ..strategies import get_benefit_strategy
-from .benefit_grant import benefit_grant as benefit_grant_service
+from .grant.service import benefit_grant as benefit_grant_service
+from .schemas import BenefitCreate, BenefitUpdate
+from .strategies import get_benefit_strategy
 
 B = TypeVar("B", bound=Benefit)
