+++ server/tests/benefit/service/test_benefit.py
@@ -6,16 +6,16 @@ from pytest_mock import MockerFixture
 
 from polar.auth.models import AuthSubject
 from polar.authz.service import Authz
+from polar.benefit.grant.service import BenefitGrantService
 from polar.benefit.schemas import (
     BenefitCustomCreate,
     BenefitCustomCreateProperties,
     BenefitCustomUpdate,
 )
-from polar.benefit.service.benefit import benefit as benefit_service
-from polar.benefit.service.benefit import (  # type: ignore[attr-defined]
+from polar.benefit.service import benefit as benefit_service
+from polar.benefit.service import (  # type: ignore[attr-defined]
     benefit_grant_service,
 )
-from polar.benefit.service.benefit_grant import BenefitGrantService
 from polar.benefit.strategies import (
     BenefitPropertiesValidationError,
     BenefitServiceProtocol,
@@ -346,7 +346,7 @@ class TestUserCreate:
                 }
             ]
         )
-        mock = mocker.patch("polar.benefit.service.benefit.get_benefit_strategy")
+        mock = mocker.patch("polar.benefit.service.get_benefit_strategy")
         mock.return_value = service_mock
 
         create_schema = BenefitCustomCreate(
