+++ server/tests/benefit/test_tasks.py
@@ -4,9 +4,7 @@ import pytest
 from arq import Retry
 from pytest_mock import MockerFixture
 
-from polar.benefit.service.benefit_grant import (
-    BenefitGrantService,
-)
+from polar.benefit.grant.service import BenefitGrantService
 from polar.benefit.strategies import BenefitRetriableError
 from polar.benefit.tasks import (  # type: ignore[attr-defined]
     BenefitDoesNotExist,
